Clean up
Removing comments, some cleanup, and moving the startup code into Program.cs
