+++ Constants/DirectoryPaths.cs
@@ -1,6 +1,4 @@
-using System;
 using System.IO;
-using System.Runtime.InteropServices;
 
 namespace Icarus.Constants;
