+++ Controllers/v1/BaseController.cs
@@ -4,8 +4,6 @@ using System.Linq;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 
-using Icarus.Controllers.Managers;
-
 
 namespace Icarus.Controllers.V1;
 
@@ -39,16 +37,5 @@ public class BaseController : ControllerBase
         return token;
     } 
 
-    /**
-    [ApiExplorerSettings(IgnoreApi = true)]
-    [Obsolete("Asymmetric key signing for tokens have been deprecated")]
-    protected bool IsTokenValid(string scope)
-    {
-        var token = ParseBearerTokenFromHeader();
-        var tokMgr = new TokenManager(_config);
-
-        return tokMgr.IsTokenValid(scope, token);
-    }
-    */
     #endregion
 }
